fix feed sort top labels
- fix sort.top.Week showing sort.top.Hour
