+++ src/hooks/contextMenu/useFeedSortOptions.ts
@@ -1,7 +1,8 @@
 import { useMemo } from "react";
 import { useTranslation } from "react-i18next";
-import { ContextMenuOption } from "../../types/ContextMenuOptions";
-import { ICON_MAP } from "../../constants/IconMap";
+import { useTopSortOptions } from "@src/hooks/contextMenu/useTopSortOptions";
+import { ContextMenuOption } from "@src/types/ContextMenuOptions";
+import { ICON_MAP } from "@src/constants/IconMap";
 
 export const useFeedSortOptions = () => {
   const { t } = useTranslation();
@@ -44,56 +45,6 @@ export const useFeedSortOptions = () => {
         key: "Old",
         title: t("sort.Old"),
         icon: ICON_MAP.HOURGLASS,
-      },
-    ],
-    [t]
-  );
-};
-
-export const useTopSortOptions = () => {
-  const { t } = useTranslation();
-
-  return useMemo<ContextMenuOption[]>(
-    () => [
-      {
-        key: "TopHour",
-        title: t("sort.top.Hour"),
-        icon: ICON_MAP.CLOCK,
-      },
-      {
-        key: "TopSixHour",
-        title: t("sort.top.SixHours"),
-        icon: ICON_MAP.CLOCK,
-      },
-      {
-        key: "TopTwelveHour",
-        title: t("sort.top.TwelveHours"),
-        icon: ICON_MAP.CLOCK,
-      },
-      {
-        key: "TopDay",
-        title: t("sort.top.Day"),
-        icon: ICON_MAP.CLOCK,
-      },
-      {
-        key: "TopWeek",
-        title: t("sort.top.Hour"),
-        icon: ICON_MAP.CALENDAR,
-      },
-      {
-        key: "TopMonth",
-        title: t("sort.top.Month"),
-        icon: ICON_MAP.CALENDAR,
-      },
-      {
-        key: "TopYear",
-        title: t("sort.top.Year"),
-        icon: ICON_MAP.CALENDAR,
-      },
-      {
-        key: "TopAll",
-        title: t("sort.top.AllTime"),
-        icon: ICON_MAP.TROPHY,
       },
     ],
     [t]
